chore: Replaced our custom file watcher for our TW plugin (#1735)
