+++ sites/skeleton.dev/vite.config.ts
@@ -1,5 +1,6 @@
 import { sveltekit } from '@sveltejs/kit/vite';
 import type { UserConfig } from 'vite';
+import skeletonPluginWatcher from './vite-plugin-skeleton-plugin-watcher';
 import sveld from './vite-plugin-sveld';
 
 // Import package.json version
@@ -8,7 +9,7 @@ const json = readFileSync('../../packages/skeleton/package.json', 'utf8');
 const pkg = JSON.parse(json);
 
 const config: UserConfig = {
-	plugins: [sveltekit(), sveld()],
+	plugins: [sveltekit(), sveld(), skeletonPluginWatcher()],
 	define: {
 		__PACKAGE__: pkg
 	}
